Move decoding code to seperate file
